add statistics module and docs
