+++ docs/api/statistics.rst
@@ -1,0 +1,3 @@
+******************
+Statistics Modules
+******************
